refactor: 상수 폴더로 분리
